Bind dbaccessor to loopback for k8s
The prior change bootstrapped dqlite to a cloud-local address where
possible. This included k8s setup, although it was never required
because CAAS allowed HA. Unfortunately, this becomes a problem, if
you delete the pod and it becomes rescheduled. When comming back it
can't rebind to the address, as it's not the same one.

To fix the problem, we prefer loopback address for CAAS deployments.
The node manager now gains that capability to know if the nodes
should perfer loopback or not. Then in the dbaccessor worker, we
can the ask the node manager for the information.
As the node manager is setup in the manifold for the dbaccessor
worker, we can then employ the IAAS vs CAAS manifolds to provide
the different functionality. This is a leaf from the 4.0 branch,
where we're trying to remove the `if isCAAS`-isms.