Remove gridfs / blobstore package
The blobstore package is no longer required, as we're using dqlite
for metadata storage and file/s3 for blob storage.

The changeset removes the dependency on state altogether and just
allows only file or s3 backends.

Unfortunately, we have a slight issue where we can't use dqlite
in the state package. So for now I've implemented an inmemory testing
metadata service. This should be removed once dqlite is everywhere.