Remove the need for ha-space on the peergrouper
This patch is a step forward in removing the legacy state spaces/subnets
domain.
Particularly, here we get rid of the need for querying the spaces when
updating the member's address in the peergrouper. The case where this
has an impact is when more than one address is available for a member:
in this case we don't request the user to set the ha-space config and
select the address that corresponds to that space, we simply select one
address from the list without failing.