Implement ModelManager facade version 10
Due to the presence of default-series in facade 9, 4.0 can only support
facade 10 model manager.

However, to allow model migrations from 3.4 -> 4.0, this means 3.4 must
support facade version 10 so the controllers can talk to eachother.

Below is the PR where 4.0 ModelManager was revved to 10
https://github.com/juju/juju/pull/16008

All that is required to support version 10 is to ensure that
default-series and default-base are not present in ModelInfo or
ModelSummary. Since these options have 'omitempty', rpc params will be
compatible so long as we ensure this field is empty