Move the tmp file location for the objectstore
When putting files into the objectstore, we first need to place them
into a temporary location.

This is for two reasons:

 - Checksum of the file to ensure it matches the expectation
 - Prevent reads whilst streaming the new file into location and
   rename (atomic on same filesystems - this part is key) to the desired
   location.

As each namespace is self managed, we need to have a better location
to ensure clean up. Unfortunately on container based workloads (k8s),
we aren't guaranteed to be on the same filesystem (different mounting
strategies). Moving the tmp to the same location as the rootPath then
allows us to have a clean location.

Once the worker bounces, we'll also attempt to clean up any files
that might be left. The pruner worker (new PR comming) will run at
a given interval to clean up any temp files left as well.