Pass all the availability zones when discovering subnets in clustered lxd
At subnet discovery on the lxd provider we only got the current server
info and use its name as availability zone, regardless of being in a
clustered environment or not.

With this patch we now fill the availability zones correctly from the
list of cluster members and correctly construct the returned subnets.