Provide controller config rules
The following allows the defining of controller config rules to prevent
certain keys from moving to other values without first being validated.

The driver for this change, is to enable a object store from moving to
s3 from file backed storage. To do that gracefully, we need to check
what the current value is and the new value. If both values are correct
the update can proceed.

This provides a nice way of ensuring the controller config updates are
atomic, that they do happen in the service layer and that we're not
exposing too much information at each layer.