Update ubuntu minimal image filtering on OCI
On the patch https://github.com/juju/juju/pull/16859 we limit the
selection of ubuntu minimal images on OCI by simply returning an error
when the "Minimal" keyword is either on the display name or in the
Operating System Version.
This patch adds a new field to the Image struct for OCI thus identifying
minimal images. For the moment we filter all images that have the
boolean IsMinimal to true, but later we might update this behaviour to
filter only some of the minimal images or non at all.