Hides state based sql errors.
After discussions within the Juju team we decided that we would like a
mechanism to hide the sql error types that bubble up from the domain
state layer.

The primary purpose of this is stop code that could potentially check
for these types outside of the state layer.

With this change maskError now implements Is and As from std errors and
checks to see if the caller is trying to establish an error type that is
for one of our sql drivers and returns false.

maskError will not try and hide any other error types. This commit just
introduces the change but does not attempt any inteogration at the
moment.