Add AccessType and ValidateAccessForAccessType.
Consts for the AccessType are easier to handle and reason than
using strings everywhere. Include a validate method to ensure
the correct ones are used.

ValidateAccessForAccessType is a helper for the new Permission domain
service layer. Validate that the access and access type combination are
allowed, without knowning the access type first.