Service layer for the new permission domain.
The service layer represents methods needed for various targets
requiring user permissions to create, read, delete and update them.

Previously all of the targets had individual methods to do similar
things for their type. This is not needed in the dqlite world. An
initial attempt to replace the various types of Read methods has been
made. This will evolve as the state layer is wired up.