Fix test failures resulting from forward merge
worker_test.go setup now included 'expectTrackedDB...' to be called.
This needed to be added to new tests introduced in the forward merge

A bit of work here needed to get worker death working. In the end, I
needed to convert the new tests over, closing the db channel and
'CleanKill'-ing in a defer

worker_namespace_tests were not aware of the new 'IsLoopbackPreferred'
method on the interface. Add some expect statements to account. Also
write some new tests for the different LoopbackPreferred values.

In doing so, extract some repeated code into a helper function,
'startWorker'