Change AccessType (string) to ObjectType (int).
Part of updating the permission service layer to not use tags introduced
an ID which represents what the tag was. Tags should only be used over the
wire.

Having the ObjectType as an int, mean it can be matched against the object
type table in the new dqlite schema.

Updated the object type and Access value validation to be against the ID,
which contains the type. ID does its own validation.