Suppress klog log messages
The klog log messages are noisy and fill up the logs really easily.
It then becomes hard to workout what is signal and what is noise.
As log messages are probably important, it's worth just suppressing
the message rather than blanket ignoring it. As the klog might be
dumping a lot of messages for large controllers, I've ensured that
we only do the request once we have a match.

The only match we have atm is the `Use tokens from the TokenRequest`.
We can always add more if we require.