Copy remaining charms handler tests to s3 charm handler
Also make some minor tweaks to the handler

When the s3 object handler was first implemented for charms, it was a
very light weight shim around the old handler. So the test suite was
made up of a small number of tests copied over from the legacy handler,
converted over to the new format

Work is being done (and is in progress) to decouple the handlers, so the
old one can be removed. So far, however, the tests have remained
relatively unchanged.

Convert over the remaining relevant tests. NOTE: Some tests were not
copied, and represent the different capabilities of the handlers. We
cannot, for instance, request a specific file from a charm with the new
handler

Make some minor cleanness tweaks + fixes to the objects handler as well,
whilst we're here