Lint against exported and unexported docs
This is the first round of linter rules added to the code base to
improve the code base at large.

The following linter rules that I added:

 - unexported-naming
 - exported
 - atomic
 - defer

I really wanted to add "loop" argument to defer, but unfortunately,
the fallout requires a lot more time and effort to undo some of the
problems.