Ensure we close client
The cross model remote secrets client would call to a remote model
to access secret information. Unfortunately the client was never
closed once it had made that request. Thus it would keep the connection
alive and all of it's resources alive preventing the clean up. As the
client is actively working, the garbage collector wouldn't be able to
clean up the resource (the monitor in the apiclient prevented that) and
it would steadly grow, based on more usage.

This fix is a relatively light touch, but I can't help but think the
better approach would be to have a pool of connections to other models,
keyed on the sourceUUID.