Implements the objectstore draining
The objectstore draining is implemented in a goroutine. The concept
is to allow the worker to spin up and drain files from the file
system if it has any. This will be repeated every restart of the
controller. We might have to consider optimising this, but every
model will need a flag for if draining is completed. Something
to discuss in the future. For now it should be fairly light on the
filesystem. The absence of a file assumes that it was already uploaded.

All requests are serialized with the main loop to place nicely with
other object stores. If we're processing files whilst draining, we
have the option to use the file backed accessor. If that fails then
it will attempt to use hash to locate the file. This prevents
blocking the controller from processing all files if it has them
locally.

This can only be accessed from controller-0. Any other controller
will skip the draining process completely.