Allow draining
We only allow draining on the bootstrapped controller. You can't
have multiple controllers if you've only got file backed object
storage. To prevent the draining process running on multiple machines,
we can check to see if it's a bootstrapped controller based on
the bootstrap params file.

We could add more checks to ensure it's valid yaml if the case
arises, but for now I think this should be sufficent.