Add test for unsetting annotations for models in annotation state layer
- Rename error type in annotation domain
- Simplify referential column in annotation tables

Which allows us to have only one variable (table name) across the
different annotation tables, which in turn simplifies the query
generation at the state layer.

- Simplify unsetting in annotations state layer

With this change, for SetAnnotations, we remove all the annotations
and then insert the given ones, so unsetting an annotation is implicit
(i.e. just don't provide in the input map).

Note that this has a small consequence of having to provide the whole
set of annotations for an entity even if one of them is to be updated
in a SetAnnotations call.

- Remove redundant query utility from internal database statement.

Because SQLair now has the slice support, see comment
https://github.com/juju/juju/pull/16896#discussion_r1488646870

- Return nil instead of initialized empty list in annotations state layer
- Check against sqlair.ErrNoRows instead of zero length slice in annotations state layer