Remove all other uuid oocurrances.
This removes all the utils.UUID occurances from the codebase. Thus
we can start to move our reliance on the utils package. The driving
reason for this change, is so that we can correctly own the code
without fear of changing it. This is the code that juju depends
on and we should own that without the constant churn of keeping our
dependencies uptodate. We spend way too much time on doing that and
it's becoming a hinderance when trying to deliever features.