Add object store type to agent config
For the first pass we should hard restart based on the object
type changing. Later on we can be smarter by only restarting when
we've got all the credentials valid. It's complicated if multiple
values have changed whilst the object store type has changed.
Blocking in that scenario could lead to unexpected results and
you can always bounce the process manually.