Move api host ports filtering on management space to the bootstrap worker
The function FilterHostPortsForManagementSpace which was added on a
previous patch to the new network domain is actually not needed as a
service function. We refactor things in order to only use the spaces we
already have during bootstrap in the bootstrap worker.

Note: this patch originates from the comment https://github.com/juju/juju/pull/16832#discussion_r1477339103