Migrate utils.Password to password.Password
Following on from #16913 this brings in the password functionality
to the internal package. This is an onging effort to centralize our
dependencies. Having dependencies that are scattered into different
repos, ones that aren't updated causes them to rot away. The uuid
package from #16913 isn't 100% spec uuid v4, so we will need to
refactor that to bring it up to spec. Nobody noticed that, and we
should also audit the password package as well, to ensure that it
is conformant.