Insist users provide a base for pre-deployed charms
This means we no longer need to parse a base from the series of
the charm url, which we can no longer guarentee will be present

Deploying pre-deployed charms is a very niche use case, which requires
the user to query the yaml or json output of "juju status" to find the
charm url. So a new breaking requirement like this is not a concern

A question was raised around how this would interact with bundles.
Particularly export bundle. However, this can be dismissed because
bundles and local charms are really compatible. It's not really possible
to deploy a bundle that makes reference to local charms

When, however, local charms are included in a model when export-bundle
is run, the "charm" field is filled in with the local charm url, and
bases/default-base is always included