Compute the 256 hash when moving to s3
Unfortunately, when we move from the filesystem to s3, we need to
recalculate the hash from sha512-384 to sha256. This ensure that
what we're actually uploading matches the expected hash. What we
don't want is to upload the file, only to find out it's not the
same hash in the future. In this scenario there wouldn't be anything
we can do.

We did talk about moving the whole of juju to sha256, but there is
a lot of work required to do that. More work than we're willing to
invest in at this time.