Adds bootstrapping of user with no auth.
It is sometimes needed that we must register a user in the Juju database
with no authentication information. This could be in cases where the
Juju user comes from another system or cases where we just need the user
for testing to create ownership links.

This change adds a new function to bootstrap for adding a user with no
auth. It also updates the bootstrap functions to validate usernames that
are being passed in as they are created outside of the bootstrap process
and may violate the buisness logic we have.

Change also creates the admin user in the apiserver suite for testing.