Use user service for local user authenticator
We know directly if the auth tag is a user, we don't need to jump
through hoops to find out if it's a tag from state. Instead, we'll
directly speak to the user service. This removes the need for
find entity (one down, many to go!).

The code was re-arranged to make it a bit easier to grok. I'm of
the opinion, that we should probably just yank all the auth out
into one location. That might be a task for another day.

More work is require, probably requiring #16935 to land first, as
the user domain state package has bugs.