Removal of metrics code
The following removes all of the metrics code. This includes the
SLA/Metric found on applications, models and charms. This currently
doesn't compile because we need to bring the charm repo into this
repository as charms will no longer be metered.

This change is because we're not collecting metrics via romulus or
omnibus. This code would have had to migrated to dqlite. Removing
this code just removes the necessity to perform that work.