Ensure we set the user last login when checking creds
We should work out if this is the right place to update the last login,
as we might not be authorized to view the content in question. For
now I'll leave it but bring it up for discussion.

I also refactored the check creds to only test for exact tags. This
should make it explicit about what we're checking credentials for.