Move spaces reload for initial models to bootstrap worker
Currently we make sure that the spaces (and subnets) are reloaded again
from the provider for the initial model if the user has provided any.
Since all the spaces steps are being moved from the agent bootstrap to
the bootstrap worker, this is a previous step to moving the space
reloading from provider completely.

It must be noted that the way of reloading spaces for the initial model
has changed on this patch: instead of realoding everything from the
provider and then filtering and recreating the subnets from the fan
config, we simply copy all the spaces and subnets (except for the HA and
management spaces) from the controller model to the initial model. We
can do this because at bootstrap we are sure that only the provider
spaces have been inserted into the database.