Model facade context
Certain facades require access to another models database. Modelmanager
allows the selection of a model by tag, which means requesting the
database of another model. We don't have a way to do that currently
and we have discussed in the past about forcing API requests for things
like this. Unfortunately, this seems rather heavy handed.

Instead the concept here is to ask for model level context at the
facade registry level. Thus if you really require model level access,
then it's opt-in. The signature of the register method changes and
you'll get a different context.

Behind the scenes everything is a model context, and we're only
restricting what you can see via the interface returned. We're not
trying to block developers from accessing methods on an interface,
but to be able to flag in code reviews or some such, if a certain
context is actually required. This could be expanded to in the
future to register smaller interface selection, thus reducing how
much dependencies the register method want's to deal with.