Duplicate jujud, split controller out.
Towards building jujud without musl, this begins an effort to break the
controller out of the jujud binary into it's own jujud-controller
binary. This new binary is effectively what jujud is before this commit,
with the jujud binary now losing the ability to be a controller.

In the meantime, since the mechanics for distributing jujud-controller
do not exist, this commit also changes the Makefile jujud and
jujud-controller targets to pick the jujud-controller binary as jujud.