Use socket instead of SIGHUP to trigger controller config reload.
The controllerconfigagent used to listen for a SIGHUP. If it recived one
it would reload all the workers watching the controller config. This
does not work on kuberenetes since there may be multiple controllers in
diferent enviroments that the SIGHUP does not reach.

The solution to this problem is to add a new socket and change the
controllerconfigagent to listen out on that socket for a reload
request. When it recives the reload request it can take the same action
as before.