Agent authenticator factory
The following allows to create an authenticator that closes over
a given state. This is so we can swap between the system state and
the model state.

I've also created an interface between stateautthenticator and
authentication. This splits up the concerns and dependencies between
the two.

We're not passing all the tests yet, but we're a step closer.