Environs at the service level
Environs are required for performing various tasks. Previously,
environs were constructed adhoc at the apiserver level. This meant
to potentially messy creation. The proposal here is to inject an
environ factory at the service level, so that any given service
could just request one at start up. If we're pushing all in for moving
logic into the services then a environ is a requirement. This felt
like an elegant solution, except for one part. It won't work becuase
of circular dependency on state. The state factory requires
the service factory for policies (that need to be removed at some
point) and the service factory requires state for model and model
config information. We obviously can't have a cyclic dependency,
so this is blocked until we remove policies.
I was actually unware (forgotten) of states dependency on the service
factory, which means you need to have a service factory up and running
before you can have a state. This feels wrong. We should remove that
dependency asap.