Replace Update with UpsertPermission.
Further investigate revealed that all of the target types have the same
sort of method in their facades to grant/revoke permissions and maybe
add the user. Move this functionality down into the domain. Renamed
Update to indicate that a user may be created.

Don't pass service structures to state. Use the state versions instead.