Move charm repo into juju/charm package
The charm repo is locked stepped with the current juju package. For
example:

  v11: juju-3.3
  v12: juju-3.4
  v13: juju-4.0

This causes problems as v13 has now blocked any new releases to
juju-3.5. Considering that every major tag is a new version of juju,
we might as face the fact that the charm repo should live inside of
juju/juju.

We've discussed this internally and agree that this is the best path
forward. The charm package should live inside the internal package, but
until we can get the terraform provider off from the charm package,
then it can live in the root.