Aligns start up probe failure with that of api.
We introduced start up probes to the Juju controller pod to force the
readiness and liveness checks to not run until we have confirmed that
the Kubernetes startup probes have succeeded.

However due to the fact that their exists a chicken and egg situation in
which the mongo container requires the controller container to
initialise some data it was possible for the mongo container to enter a
failure state waiting for the controller.

What this fix is doing is aligning their startup probes for the moment
so they at least are given equal startup time.

However the correct fix long term is move this bootstrapping activity to
an init container and decouple the two containers from each other so
they can startup independently.

Addresses lp 2054930