Remove assign clean to machines
Assign clean machines is the start of removing the policy implementation
from within state. The concept of assigning to a clean machine or
event a clean empty machine thought of machines as pets rather than
cattle. We should be just creating new machines and decommissioning
the machines once they're done.

This removes the distribute unit code from state, along with the
tests.