Add max values to NumProvisionWorkers and NumContainerProvisionWorkers.
Theoretically any user with model access could bring the controller to a
standstill by setting the number of provisioner workers to an extremely
large value. Set a sane value as a cap.

The values are larger than the default but not back by solid data today.
We do know that 16 workers for containers will cause issues with LXD on
some machines.

In the future we should allow the controller config to define what the
max values should be. Today controller config has no linkage with model
config to enable this. A question would be whether existing model config
should be changed if the max is changed.