Wire up activation key to register
To ensure we're only allowing one activation key for a user, we
need to remove the old state code which sets the key. In the old
code that was called a secret. We want to remove that terminology,
as it's confusing with the new ongoing secret work.

The code pushes the nacl secret box into the service layers, so
we can live closer to state. The API only then unpacks the transport
layers and hands it off.