Bump Pebble version to v1.9.0
This Pebble version adds the following since v1.7.1:

- canonical/pebble#368: replace "select=all" with "users=all". Strictly
speaking this is a breaking change, however, it was a design mistake
and we want to fix it as soon as possible. It's almost certainly that
no one is using this feature yet.
- canonical/pebble#369: avoid acquiring state lock in health check
endpoint. This fixes (or is one aspect of the fix for) the issue
described in https://bugs.launchpad.net/juju/+bug/2052517, so that the
GET /v1/health endpoint returns much quicker even when under load.

In addition, this has the side effect of bumping up a few of the
golang.org/x packages, as Pebble depends on higher versions.