Fix --to placement directive
This resolves https://bugs.launchpad.net/juju/+bug/2054375

The charmhub api does not expect a risk to be present in OS base channels.
However, under some circumstances (importantly, using the --to placement
directive), a risk is included. Filter these out

Add an integration test to cover this case

Also, as a flyby, remove the ChannelTrack function from core charm. This
function is confusing and encourages misuse. It was only used twice, on
both occasions to filter the risk out of an OS base. However, the
function lived in the charm package, and it's name/description imply
it's to be used for charm channels.

A few options were considered for what to do instead. In the end, I made
the decision to just remove it, and in the places it was used to parse a
base channel and just read the Track.

This turns out to be exactly what a 'BaseChannelTrack' function would
do, and not actually add any complexity to the code. In my opinion, a
function/method to do this for us is not a helpful abstraction, since
dropping it actually increases readability.

Confusion around OS channels and charm channels is avoided because call
a concrete base.ParseChannel function, and deal with the concrete
base.Channel struct. This would be hidden if we use a 'BaseChannelTrack'
function.

This isn't perfect, however. There is still a bit of overlap between
Juju-base-channel and charmhub-base-channel, which aren't quite always
the same. We would now need to support channels with track "all" for
instance. This was the case beforehand charm channels, so this is an
improvement.