Move downloader and repository to internal charm package
The following no longer need to be core. They're implementation
details of core/charm/Repository interface. This is a long road
to bringing in the charm package into juju/juju. Unfortunately,
we've made a mess when it comes to bases and channels. We'll need
to unpick that before we can bring in the repo, ideally before 4.0
release.