Refactor policy instance prechcker
This is an ongoing effort to lift the instance prechecker out of
the policy and make it a dependency of every method/function. Thus
we can then just pass it in when required.

I'm wishfully thinking that returning a noop instance prechecker will
cause most if not all tests to pass. We don't care that much about
the state package, but we do want the API server or any additional
worker to continue on.

The state package should be up to date, we'll have to do this for
every single method on the policy.