Adds a view to the ddl for credentials.
Currently we primarily deal with credentials using the natural key of
cloud name, owner name and credential name. It is however required that
we flick back to the uuid of the credential when we need to be exact.

This commit adds a view to the table so that this data can be easily
accessed.