GOCOVERDIR causes tests to fail
Some tests actually invoke the go tooling to run a test, which causes
a failure as we now set juju to run with go coverage. We want to run
code coverage so that we can push that information to tiobe.

Unfortunately, GOCOVERDIR outputs a warning if the environment variable
is not set. This warning is then caught by the strict equality output
we're looking for on stdout/stderr.

Note: these tests should be deprecated going forward.