Make sure subnets are updated when a NIC gets an address on manual cloud
On manual clouds, when a new link layer device is added, the subnets are
processed so that they include the new device's address. This works fine
except for the case when a NIC doesn't have an address assigned at the
time of provision but gets one later.

This patch therefore makes sure that the subnets are updated both when a
new NIC is detected and also when a new address for an existing NIC is
detected.