Wire up the dqlite storage pools. Previously, we used a pool manager and mongo settings collection - pool manager is completely deleted.
To allow the business logic for storage in state to still operate, the
state policy is updated to allow the storage service to be wired in.
This provides access to named storage pools so at deploy time, any
storage directives can be used by looking up the pool by name.

There's a chicken and egg issue, between mongo and dqlite, so model
import and hence migration will fail if there are volumes using a custom
storage pool. We'll need to live with that for now.