Refactor policy config schema source
Removes the ProviderConfigSchemaSource from state policy. Interestingly
this didn't require state and so could have been lifted out when the
cloud service was added.
The ongoing effort is to remove the concept of policy from state and free
us from the cyclic dependency within the engine. Only a few locations
require the config schema source and it was clear the abstraction for this
didn't hold it's weight.

The interface now becomes:

```
type Policy interface {
	// ConfigValidator returns a config.Validator or an error.
	ConfigValidator() (config.Validator, error)

	// ConstraintsValidator returns a constraints.Validator or an error.
	ConstraintsValidator(envcontext.ProviderCallContext) (constraints.Validator, error)

	// StorageProviderRegistry returns a storage.ProviderRegistry or an error.
	StorageProviderRegistry() (storage.ProviderRegistry, error)
}
```

The next set of changes will attempt to remove the config and constraints
validator from the policy at once. A lot of tests will be broken, so it'll
be a case of working through those to be able to land this.