Remove dependancy on legacy state for application bindings
This patch removes the internal dependency on spaces in the legacy
mongodb state regarding endpoint bindings. We do this by injecting the
list of all available spaces for lookup, which is needed for validating
the bindings.

This change requires passing this new value into the EndpointBindings()
method which itself requires many other function signatures to be
changed in state, namely AddApplication(). This explains the huge number
of modified files and lines of code.