Remove all remaining GETs and SETs of Series in charm url
There are still a few instances in unit tests, which will need to be
removed later. However, a significant number have been removed here to
fix tests

The only exceptions beyond this are in apiserver/objects.go and
apiserver/charms.go. This is because the legacy handler still needs to
understand the concept of the series for migration.

When a source controller migrates a model, the controller asserts that
the new controller does not mutate the charm url. This means the handler
needs to be able to reconstruct the charm url entirely, including the
series.

Now that we onlyt read charm url series in one optional place, we can
drop every instance of writing the series.

This will require some thorough testing, since charm urls over the wire
before and after this change will stringify differently, which could
result in subtle bugs.

Dropping series from charm url in core/charm/charmpath.go meant we could
significantly simplify NewCharmAtPath. This code is now nicely
de-coupled from any concerns about selecting an appropriate base. This
had some ramifications, since we use this code in a fair number of
places.

One particular consequence in the client is that it now makes sense to
refactor CharmReader to call NewCharmAtPath instead of ReadCharm.