Introduces model creation to model migration.
This PR introduces model creation in DQlite as part of model migration.
We are doing this to facilitate further changes to the model DDL that
requires the removal of the modelmanager service.

To achieve this removal we need to properly create models during
migration. As well as this because of how application and machine
importing was working we also need to migrate these two entities into
the model database.

This commit also disables the current model migration tests that we
have. This is because they are currently not setup to properly handle
DQlite migration and will need to be addressed in follow up PR's.