Add Prepare to StateBase and update SQLair
Add a Prepare function to StateBase which caches sqlair statements. This
allows for statement reuse rather than having to reprepare a statement
every time it is used.

The latest update to SQLair introduced a sqlair.ErrNoRows error when
GetAll finds no results, there are quite a few places in the code where
GetAll is used and this caused some error. Checks are put in to fix
this.