Expose logging for errors
Unfortunately, there exists a bug from the dqlite driver that returns
an error code that doesn't match the expected error code, even though
the error message is correct.

I'm unable to reproduce this locally, so I'm adding additional logging
with the hope that if someone spots it in the wild they'll pass the
information on. I've set it to critical just to catch it.