Idempotent bootstrap worker
The bootstrap worker can retry if there is a failure. When switching
over the objectstore, we didn't update the errors that the bootstrap
worker can handle. If the hash already exists, we can ignore that
error and continue the flow as if the objectstore was idempotent.

We still want the error because in other scenarios we want to log
out that error. This is the best of both worlds.