Add provider service factory worker
Implements the provider service factory worker along with cloud
and credential provider services. The provider services aren't
just regular services, instead they're a subset of methods that we
offer out. If the return type of the service factory was an interface
we could just have a different interface, but we use concrete types,
so we need to re-encapsulate that logic. Offering out a full service
could give developers the wrong idea in the future, that the
service factory and provider service factory could be used
interchangably. This would be a miss-step, we want to ensure they're
different types, so that nobody acciedently depends on them.

The services themselves offer out what the full service offers, but
that doesn't always have to be the case, we could diverge as they do
satisfy other requirements.

Most of this is a clone from the service-factory worker code, so it's
reuse is heavily inspired here.