Generalize provider tracker
The provider tracker is almost identical between providers and brokers.
As there is a common overlap between then, we can use generics to give
us one worker that provides access to both. Then we can remove all the
duplicate code for both.

This is a proof of concept whilst we experience with the generic side
of things.