Thread provider service through model dep engine
Now that we've got the provider service factory, we can thread it
through the dependency engine manfiolds. We want it as a dependency
in the engine, so we can reuse the provider and broker workers outside
of the model dependency engine. If they're config only options, then
it means that we can't supply them in other locations (it's easier
to lift into the manifolds, than to extract).