Provide read-only model in provider-tracker
We can use the newly created read-only model to access the required
fields for a environ. The new provider tracker will access the data
directly from the database/service factory rather making a round
trip via the apiserver. This should elevate the amount of internal
requests we have to perform on ourselves.