Add get model
We need the model data for the environ tracker worker. I know we're
potentially renaming model_metadata to model, but for now I just need
to get a model.

We already have all the information, so I'm just getting the
information back out. As part of that, I needed to have a core Model
type. This meant moving the core/Model to api/types, as the api
client package shouldn't be using core model types (we'll be changing
this as we go, this is the first step!). Then we can expose a core
model type. All the churn then becomes, you can't bring domain into
core.