Wire up model database services and state
This pushes through the concept of a read-only model for the model
database. The data within the read-only model database is denormalized.
This is on purpose as we don't want to replicate the controller database
in the model database. The sole purpose is to have a cache of data
that require from the two. It's important to note: no data in the
read-only database should be able change. If that requirement changes,
we need to remove it from this database.

I've only added things that I think are required, but we can always
add more.