Create model table in model database whilst migrating
We need to ensure that we also create the read-only model table in
side the model database when model migrating.

As this is a two step transaction process, we need to coordinate the
model migration. The updated model migration coordinator ensures that
we have all the information during a rollback that we have in a
execution. Thus we can perform rollbacks without having a stateful
migration script.