Remove integration unit tests
The machine legacy tests start up the jujud-controller and test
various parts are working. The downside is, that we require so much
faked or stubbed infrastructure that we don't actually know if something
is broken. Normally it's actually the alternative, whereby the normal
workflow works correctly, but we're missing stub or faked code.

I've left in what I think are the important required tests. Other
tests are actually already covered via integration tests or via
actual unit tests further down the stack.